+++ src/commands/Configuration/override/delete.ts
@@ -1,10 +1,10 @@
-import { NonDigits } from "../../../../utils/Regex";
+import { NonDigits } from "../../../utils/Regex";
 import { Message, Snowflake } from "discord.js";
-import ClearanceManager from "../../../../utils/managers/ClearanceManager";
-import Command from "../../../../structures/Command";
-import SubCommand from "../../../../structures/SubCommand";
-import CommandContext from "../../../../structures/CommandContext";
-import BulbBotClient from "../../../../structures/BulbBotClient";
+import ClearanceManager from "../../../utils/managers/ClearanceManager";
+import Command from "../../../structures/Command";
+import SubCommand from "../../../structures/SubCommand";
+import CommandContext from "../../../structures/CommandContext";
+import BulbBotClient from "../../../structures/BulbBotClient";
 
 const clearanceManager: ClearanceManager = new ClearanceManager();
