+++ src/commands/Configuration/override/list.ts
@@ -1,11 +1,11 @@
 import { Message, MessageEmbed, Snowflake } from "discord.js";
-import { embedColor } from "../../../../Config";
-import * as Emotes from "../../../../emotes.json";
-import BulbBotClient from "../../../../structures/BulbBotClient";
-import Command from "../../../../structures/Command";
-import SubCommand from "../../../../structures/SubCommand";
-import CommandContext from "../../../../structures/CommandContext";
-import ClearanceManager from "../../../../utils/managers/ClearanceManager";
+import { embedColor } from "../../../Config";
+import * as Emotes from "../../../emotes.json";
+import BulbBotClient from "../../../structures/BulbBotClient";
+import Command from "../../../structures/Command";
+import SubCommand from "../../../structures/SubCommand";
+import CommandContext from "../../../structures/CommandContext";
+import ClearanceManager from "../../../utils/managers/ClearanceManager";
 
 const clearanceManager: ClearanceManager = new ClearanceManager();
